+++ CHANGELOG.md
@@ -6,3 +6,4 @@ An overview of the changes per-version for the Dugway gem.
 
 - Switch from Travis CI to GitHub Actions for project CI
 - Support Ruby 2.7.x
+- Add `ProductDrop#related_products`, which returns a `ProductsDrop` for rendering related products for a given one